+++ packages/integration/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0](https://github.com/mjancarik/merkur/compare/v0.16.2...v0.17.0) (2020-08-27)
+
+
+### Features
+
+* 🎸 base integration package ([8320b22](https://github.com/mjancarik/merkur/commit/8320b22a4d46cf99678f372defe8d84f4488aaa9))
+* 🎸 integration-react using base integration pkg ([a3aef85](https://github.com/mjancarik/merkur/commit/a3aef859a80ff0a884d32fc5c98f77843b55f5f0))
+
+
+
+
+
+## [0.16.2](https://github.com/mjancarik/merkur/compare/v0.16.1...v0.16.2) (2020-08-25)
+
+**Note:** Version bump only for package @merkur/integration
+
+
+
+
+
 ## [0.16.1](https://github.com/mjancarik/merkur/compare/v0.16.0...v0.16.1) (2020-08-14)
 
 **Note:** Version bump only for package @merkur/integration